blehh

+++ Module 2/arithmetic.c
@@ -1,13 +1,13 @@
 #include <stdio.h>
 int main(){
     int a = 10;
-    int b = 4;
+    float b = 4;
     int sum = a+b;
     printf("%d\n",sum);
     int sub = a-b;
     printf("%d\n", sub);
     int mul = a*b;
     printf("%d\n",mul);
-    int div = a/b;
-    printf("%d\n",div);
+    float div = a/b;
+    printf("%f\n",div);
 }
